W174y: FOUR MATCHes across the Kosaka/Social sweep -- and a CONFIRMED FLOOR falls.
  sfl_card    00258630  nd6   -> **MATCH**   <-- was on the confirmed-floor list
  k_sceneDraw 0019F8F0  nd22  -> MATCH       the > 0xff clamp form
  k_draw      001A57A0  nd52  -> MATCH       local lifetimes plus a component-wise RwV3d copy
  k_clump     001A7910  nd134 -> MATCH       D_00678C60 typed, unsigned loop bound, > 0xff clamps, interleaved RwRGBA aggregate construction/copy
  k_data      001B8710  nd36 -> 18 (344/352) and 001B8960  nd36 -> 18 (376/384)
  frFont      003AFE30  nd189 -> 78, 284 -> 272/288   typed manager/slot layout plus a HIDDEN DESTRUCTOR ARGUMENT

**sfl_card 00258630 was listed as a confirmed floor and it was wrong.** The fix was to stage the embedded card pointer and access card[-4], forcing retail's argument setup order. **The floor list is a heuristic, not a proof** -- it records where someone previously failed, not where success is impossible. Three entries have now fallen this campaign after being written off. Treat a floor as low-priority, not as forbidden, especially when a new lever arrives that nobody had when it was declared.

The  clamp form closed one function outright and contributed to another. It forces retail's  materialization and is cheap to test wherever a value is clamped to a byte range.

REGRESSION REPAIRED: the agent left fmGslCont 003B4B40 at nd15 against an nd11 baseline -- during a revert it restored a missing z assignment in the wrong ORDER relative to the vertices assignment. Swapped them back; verified nd11 restored. **Note src/fmGslCont.c is LF-ONLY** (0 CRLF / 1,056 LF), a third such file alongside btlFormation.c and itfMesManager.c -- my first repair attempt failed silently because it assumed CRLF. Splice in bytes and check the endings per file.

Honest negative: D_00960184 is already correctly typed void* (*)(u32,u32,u32) and retail's call-argument census for the k_data pair shows NO hidden or invented allocator argument. Their remaining nd18 is cached-size load SCHEDULING, not ABI width.

576/700 across the slice, no lost MATCHes, no over-window additions, no stale or NO_SYMBOL entries.

+++ src/Battle/btlTarget.c
@@ -2170,7 +2170,7 @@ short FUN_002d9d70(int param_1)
   uint uVar5;
   uint uVar6;
   uint uVar7;
-  long lVar8;
+  s32 lVar8;
   undefined8 uVar9;
   ulong uVar10;
   short sVar11;
